backend/requirements.txt
flask
flask-jwt-extended
flask-pymongo
flask-bcrypt
python-dotenv
itsdangerous
